Made test reusable
